% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedfile.R
\name{sharedfile_path}
\alias{sharedfile_path}
\title{Use shared files from an orderly task}
\usage{
sharedfile_path(files, from = NULL)
}
\arguments{
\item{files}{A character vector of filenames}

\item{from}{Optional string, required if you have more than one
folder configured.}
}
\value{
The absolute path of the requested files, in the same
order as \code{files}
}
\description{
Return a path to a file from a shared file location. This function
returns the absolute path to a file (or directory) within one of
your collections of shared resources.
}
